Move getWordsOfDict to domain-logic
